Added forecast for other days

+++ app/src/main/java/hr/dtakac/prognoza/forecast/uimodel/TodayUiModel.kt
@@ -4,7 +4,7 @@ import java.time.ZonedDateTime
 
 data class TodayUiModel(
     val dateTime: ZonedDateTime,
-    val currentTemperature: Short,
+    val currentTemperature: Int,
     val weatherIcon: WeatherIcon,
     val nextHours: List<HourUiModel>
 )
